backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
aiohttp>=3.9.0
azure-cognitiveservices-speech>=1.35.0
pyodbc>=4.0.39
azure-storage-blob>=12.19.0
applicationinsights>=0.11.10
